(송인석) exp컨트롤러 주석 수정 및 oneAnimalExpList css,부트스트랩 적용 페이징 작업 진행
expAnimalList.jsp
	- 이미지 크기 조절 및 얼럿창 설정(리스트 확인후 돌려보냄)

ExpController.java
	- OneAnimalExpInfo메서드수정(받아오는 값 추가 및 페이징 작업진행)
	- 나머지 메서드수정(주석 변경)

ExpDao.java
	- selectExpAnimalTotalCount메서드추가(해당 체험동물 체험 기록 총 카운터 확인)
	- selectOneAnimalExpInfo메서드수정(페이징 작업진행)

expJournalList.jsp
	- 얼럿창 설정(리스트 확인후 돌려보냄)

expList.jsp
	- 얼럿창 설정(리스트 확인후 돌려보냄)

ExpMapper.xml
	- selectExpAnimalTotalCount추가(해당 체험동물 체험 기록 총 카운터 확인)
	- selectOneAnimalExpInfo수정(페이징 작업 진행)

ExpService.java
	- selectOneAnimalExpInfo메서드수정(리턴값 수정 및 페이징 작업 진행)

expShelterInfo.jsp
	- 분기코드 수정

expShelterList.jsp
	- 얼럿창 설정(리스트 확인후 돌려보냄)

oneAnimalExpList.jsp
	- css및 부트스트랩 적용, 페이징 작업진행

+++ teamProject/src/main/java/com/cafe24/iso159/ExpController.java
@@ -33,18 +33,30 @@ public class ExpController {
 	private ExpService expService;
 	private static final Logger logger = LoggerFactory.getLogger(ExpController.class);
 	
-	//체험돌물 보여줄때 해당동물 이미 진행한 체험 기록 보여줌
+	//체험돌물 보여줄때 해당동물 이미 진행한 체험 기록 보여줌 /experience/oneAnimalExpList 을 GET 방식으로 호출할때 발생
 	@RequestMapping(value = "/experience/oneAnimalExpList", method = RequestMethod.GET)
-	public String OneAnimalExpInfo(Model model,@RequestParam(value="animalCode") String animalCode) {
+	public String OneAnimalExpInfo(Model model,@RequestParam(value="animalCode") String animalCode
+							,@RequestParam(value="currentPage" ,defaultValue="1" ,required=false) int currentPage
+							,@RequestParam(value="rowPerPage" ,defaultValue="10" ,required=false) int rowPerPage) {
 		logger.debug("ExpController 호출 {OneAnimalExpInfo.get}.");
 		logger.debug("OneAnimalExpInfo().get 메서드 animalCode is {}",animalCode);
-		List<ExpAndExpJournal> expAndAnimalAndExpJournal = expService.selectOneAnimalExpInfo(animalCode);
-		logger.debug("OneAnimalExpInfo().get 메서드 expAndAnimalAndExpJournal is {}",expAndAnimalAndExpJournal);
-		model.addAttribute("expAndAnimalAndExpJournal", expAndAnimalAndExpJournal);
+		logger.debug("OneAnimalExpInfo().get 메서드 currentPage is {}",currentPage);
+		logger.debug("OneAnimalExpInfo().get 메서드 rowPerPage is {}",rowPerPage);
+		//가져온 animalCode,currentPage,rowPerPage 3개 map 에 담아서 메서드 실행
+		Map<String, Object> map = new HashMap<String, Object>();
+		map.put("currentPage", currentPage);
+		map.put("rowPerPage", rowPerPage);
+		map.put("animalCode", animalCode);
+		Map<String, Object> returnMap = expService.selectOneAnimalExpInfo(map);
+		logger.debug("OneAnimalExpInfo().get 메서드 returnMap is {}",returnMap);
+		//담겨저온 값 확인
+		logger.debug("ExpOneList().get 메서드 returnMap is {}",returnMap);
+		model.addAttribute("returnMap", returnMap);
+		model.addAttribute("animalCode", animalCode);
 		return "/experience/oneAnimalExpList";
 	}
 	
-	//체험 하기 클릭시 체험 가능 동물 내용 띄워줌 
+	//체험 하기 클릭시 체험 가능 동물 내용 띄워줌 /experience/expAnimalList 을 GET 방식으로 호출할때 발생
 	@RequestMapping(value = "/experience/expAnimalList", method = RequestMethod.GET)
 	public String expAnimalList(Model model) {
 		logger.debug("ExpController 호출 {expAnimalList.get}.");
@@ -54,20 +66,34 @@ public class ExpController {
 		return "/experience/expAnimalList";
 	}
 	
-	//해당번호 체험일지 수정 내용 조회
+	//해당번호 체험일지 수정 내용 조회 /experience/expJournalModify 을 POST 방식으로 호출할때 발생
 	@RequestMapping(value = "/experience/expJournalModify", method = RequestMethod.POST)
-	public String expJournalModify(ExpJournal expJournal) {
+	public String expJournalModify(ExpJournal expJournal,HttpSession session) {
 		logger.debug("ExpController 호출 {expJournalModify.post}.");
+		//세션에서 아이디 값 확인
+		String loginId = (String)session.getAttribute("loginId");
+		logger.debug("expJournalModify().post 메서드 loginId is {}",loginId);
+		//로그인 상태 확인
+		if(loginId==null) {
+			return "redirect:/member/login";
+		}
 		logger.debug("expJournalModify().get 메서드 expJournal is {}",expJournal);
 		//뷰에서 가져온 정보로 updateExpJournal메서드 실행해서 업데이트
 		expService.updateExpJournal(expJournal);
 		return "redirect:/experience/expList";
 	}
 	
-	//체험자 체험일지 수정
+	//체험자 체험일지 수정 /experience/expJournalModify 을 get 방식으로 호출할때 발생
 	@RequestMapping(value = "/experience/expJournalModify", method = RequestMethod.GET)
-	public String expJournalModify(Model model,@RequestParam(value="expJournalCode") String expJournalCode) {
+	public String expJournalModify(Model model,@RequestParam(value="expJournalCode") String expJournalCode,HttpSession session) {
 		logger.debug("ExpController 호출 {expJournalModify.get}.");
+		//세션에서 아이디 값 확인
+		String loginId = (String)session.getAttribute("loginId");
+		logger.debug("expJournalModify().post 메서드 loginId is {}",loginId);
+		//로그인 상태 확인
+		if(loginId==null) {
+			return "redirect:/member/login";
+		}
 		logger.debug("expJournalModify().get 메서드 expJournalCode is {}",expJournalCode);
 		//뷰로 값을 넘김
 		model.addAttribute("expJournalCode", expJournalCode);
@@ -77,10 +103,17 @@ public class ExpController {
 		return "/experience/expJournalModify";
 	}
 	
-	//해당 체험자 체험일지 리스트
+	//해당 체험자 체험일지 리스트 /experience/expJournalList 을 get 방식으로 호출할때 발생
 	@RequestMapping(value = "/experience/expJournalList", method = RequestMethod.GET)
-	public String ExpJournalList(Model model,@RequestParam(value="expCode") String expCode) {
+	public String ExpJournalList(Model model,@RequestParam(value="expCode") String expCode,HttpSession session) {
 		logger.debug("ExpController 호출 {ExpJournalList.get}.");
+		//세션에서 아이디 값 확인
+		String loginId = (String)session.getAttribute("loginId");
+		logger.debug("ExpJournalList().post 메서드 loginId is {}",loginId);
+		//로그인 상태 확인
+		if(loginId==null) {
+			return "redirect:/member/login";
+		}
 		logger.debug("ExpJournalList().get 메서드 expCode is {}",expCode);
 		//expCode로 해당 체험자가 등록한 체험일지 확인
 		List<ExpJournal> expJournal = expService.selectExpJournalList(expCode);
@@ -88,21 +121,29 @@ public class ExpController {
 		return "/experience/expJournalList";
 	}
 	
-	// 보호소 체험진행,종료시 동물 상태 변경
+	// 보호소 체험진행,종료시 동물 상태 변경 /experience/animalUpdate 을 get 방식으로 호출할때 발생
 	@RequestMapping(value = "/experience/animalUpdate", method = RequestMethod.GET)
 	public String progressionAnimalUpdate(@RequestParam(value="animalCode") String animalCode,
 										@RequestParam(value="expCode") String expCode,
 										@RequestParam(value="osCodeAnimal") String osCodeAnimal,
 										@RequestParam(value="osCodeExp") String osCodeExp,
 										@RequestParam(value="osCode") String osCode,
-										@RequestParam(value="expCost") String expCost) {
-		logger.debug("ExpController 호출 {animalUpdate.get}.");
-		logger.debug("animalUpdate().get 메서드 animalCode is {}",animalCode);
-		logger.debug("animalUpdate().get 메서드 expCode is {}",expCode);
-		logger.debug("animalUpdate().get 메서드 osCodeAnimal is {}",osCodeAnimal);
-		logger.debug("animalUpdate().get 메서드 osCodeExp is {}",osCodeExp);
-		logger.debug("animalUpdate().get 메서드 osCode is {}",osCode);
-		logger.debug("animalUpdate().get 메서드 expCost is {}",expCost);
+										@RequestParam(value="expCost") String expCost,
+										HttpSession session) {
+		logger.debug("ExpController 호출 {progressionAnimalUpdate.get}.");
+		//세션에서 아이디 값 확인
+		String loginId = (String)session.getAttribute("loginId");
+		logger.debug("progressionAnimalUpdate().post 메서드 loginId is {}",loginId);
+		//로그인 상태 확인
+		if(loginId==null) {
+			return "redirect:/member/login";
+		}
+		logger.debug("progressionAnimalUpdate().get 메서드 animalCode is {}",animalCode);
+		logger.debug("progressionAnimalUpdate().get 메서드 expCode is {}",expCode);
+		logger.debug("progressionAnimalUpdate().get 메서드 osCodeAnimal is {}",osCodeAnimal);
+		logger.debug("progressionAnimalUpdate().get 메서드 osCodeExp is {}",osCodeExp);
+		logger.debug("progressionAnimalUpdate().get 메서드 osCode is {}",osCode);
+		logger.debug("progressionAnimalUpdate().get 메서드 expCost is {}",expCost);
 		// animalCode 가져와서 체험 진행,종료시 동물상태 변경
 		Animal animal = new Animal();
 		animal.setAnimalCode(animalCode);
@@ -125,14 +166,19 @@ public class ExpController {
 		return "redirect:/experience/expShelterList";
 	}
 	
-	//해당 보호소 체험 예약 상태 진행
+	//해당 보호소 체험 예약 상태 진행 /experience/updateOsExp 을 get 방식으로 호출할때 발생
 	@RequestMapping(value = "/experience/updateOsExp", method = RequestMethod.GET)
 	public String checkOsExp(@RequestParam(value="osCodeExp") String osCodeExp,@RequestParam(value="expCode") String expCode,HttpSession session) {
 		logger.debug("ExpController 호출 {checkOsExp.get}.");
 		logger.debug("checkOsExp().get 메서드 osCodeExp is {}",osCodeExp);
 		logger.debug("checkOsExp().get 메서드 expCode is {}",expCode);
+		//세션에서 아이디 값 확인
 		String loginId = (String)session.getAttribute("loginId");
-		logger.debug("checkOsExp().get 메서드 loginId is {}",loginId);
+		logger.debug("checkOsExp().post 메서드 loginId is {}",loginId);
+		//로그인 상태 확인
+		if(loginId==null) {
+			return "redirect:/member/login";
+		}
 		Exp exp = new Exp();
 		exp.setmShelterIdAccept(loginId);
 		exp.setExpCode(expCode);
@@ -141,28 +187,41 @@ public class ExpController {
 		return "redirect:/experience/expShelterList";
 	}
 	
-	//해당 보호소 체험자 정보
+	//해당 보호소 체험자 정보 /experience/expShelterInfo 을 get 방식으로 호출할때 발생
 	@RequestMapping(value = "/experience/expShelterInfo", method = RequestMethod.GET)
 	public String expShelterInfo(Model model,@RequestParam(value="expCode") String expCode,HttpSession session) {
 		logger.debug("ExpController 호출 {expShelterInfo.get}.");
 		logger.debug("expShelterInfo().get 메서드 expCode is {}",expCode);
+		//세션에서 아이디 값 확인
+		String loginId = (String)session.getAttribute("loginId");
+		logger.debug("expShelterInfo().post 메서드 loginId is {}",loginId);
+		//로그인 상태 확인
+		if(loginId==null) {
+			return "redirect:/member/login";
+		}
 		//체험 리스트에서 체험 코드 가져와서 정보 확인
 		ExpAndAnimalAndOverallStatusAndExpPeriodAndMemberInfo selectExpOneInfo =  expService.selectExpShelterInfo(expCode);
 		logger.debug("expShelterInfo().get 메서드 selectExpOneInfo is {}",selectExpOneInfo);
 		model.addAttribute("ExpOneInfo", selectExpOneInfo);
 		//해당 보호소 체험자 리스트 에서 정보 확인할때 확인자 아이디 ,체험 상태 등록
 		Exp exp = new Exp();
-		exp.setmShelterIdCheck((String)session.getAttribute("loginId"));
-		logger.debug("expShelterInfo().get 메서드 loginId is {}",(String)session.getAttribute("loginId"));
+		exp.setmShelterIdCheck(loginId);
 		exp.setExpCode(expCode);
 		expService.updateExpmShelterIdCheck(exp);
 		return "/experience/expShelterInfo";
 	}
 	
-	// 해당 보호소 체험자 리스트
+	// 해당 보호소 체험자 리스트 /experience/expShelterList 을 get 방식으로 호출할때 발생
 	@RequestMapping(value = "/experience/expShelterList", method = RequestMethod.GET)
 	public String expshelterList(Model model,HttpSession session) {
 		logger.debug("ExpController 호출 {expshelterList.get}.");
+		//세션에서 아이디 값 확인
+		String loginId = (String)session.getAttribute("loginId");
+		logger.debug("expshelterList().post 메서드 loginId is {}",loginId);
+		//로그인 상태 확인
+		if(loginId==null) {
+			return "redirect:/member/login";
+		}
 		//blCode 값 확인
 		String loginBlCode = (String)session.getAttribute("loginBlCode");
 		logger.debug("expshelterList().get 메서드 blCode is {}",loginBlCode);
@@ -172,12 +231,19 @@ public class ExpController {
 		return "/experience/expShelterList";
 	}
 	
-	// 체험신청자 체험 삭제
+	// 체험신청자 체험 삭제 /experience/deleteExp 을 get 방식으로 호출할때 발생
 	@RequestMapping(value = "/experience/deleteExp", method = RequestMethod.GET)
-	public String deleteExpOne(@RequestParam(value="expCode") String expCode) {
+	public String deleteExpOne(@RequestParam(value="expCode") String expCode, HttpSession session) {
 		logger.debug("ExpController 호출 {deleteExpOne.get}.");
 		//넘어온 expCode 값 확인
 		logger.debug("deleteExpOne().get 메서드 expCode is {}",expCode);
+		//세션에서 아이디 값 확인
+		String loginId = (String)session.getAttribute("loginId");
+		logger.debug("deleteExpOne().post 메서드 loginId is {}",loginId);
+		//로그인 상태 확인
+		if(loginId==null) {
+			return "redirect:/member/login";
+		}
 		expService.deleteExpOne(expCode);
 		return "redirect:/experience/expList";
 	}
@@ -191,6 +257,10 @@ public class ExpController {
 		String mExpId = (String)session.getAttribute("loginId");
 		//넘어온 loginId 값 확인
 		logger.debug("ExpOneList().get 메서드 mExpId is {}",mExpId);
+		//로그인 상태 확인
+		if(mExpId==null) {
+			return "redirect:/member/login";
+		}
 		logger.debug("ExpOneList().get 메서드 currentPage is {}",currentPage);
 		logger.debug("ExpOneList().get 메서드 rowPerPage is {}",rowPerPage);
 		Map<String, Object> map = new HashMap<String, Object>();
@@ -206,7 +276,7 @@ public class ExpController {
 		return "/experience/expList";
 	}
 	
-	// 체험 일지 작성
+	// 체험 일지 작성 /experience/expJournalAdd 을 POST 방식으로 호출할때 발생
 	@RequestMapping(value = "/experience/expJournalAdd", method = RequestMethod.POST)
 	public String expJournalAdd(ExpJournal expJournal,HttpSession session) {
 		logger.debug("ExpController 호출 {expJournalAdd.post}.");
@@ -215,17 +285,28 @@ public class ExpController {
 		// loginId 를 통해서 입력 아이디 값 받음
 		String loginId = (String)session.getAttribute("loginId");
 		logger.debug("expJournalAdd().get 메서드 loginId is {}",loginId);
+		//로그인 상태 확인
+		if(loginId==null) {
+			return "redirect:/member/login";
+		}
 		expJournal.setmExpId(loginId);
 		expService.addExpJournal(expJournal);
 		return "redirect:/experience/expList";
 	}
 	
-	// 체험 일지 작성 폼으로
+	// 체험 일지 작성 폼으로 /experience/expJournalAdd 을 get 방식으로 호출할때 발생
 	@RequestMapping(value = "/experience/expJournalAdd", method = RequestMethod.GET)
-	public String expJournalAdd(Model model,@RequestParam(value="expCode") String expCode) {
+	public String expJournalAdd(Model model,@RequestParam(value="expCode") String expCode, HttpSession session) {
 		logger.debug("ExpController 호출 {expJournalAdd.get}.");
 		//넘어온 expCode 값 확인
 		logger.debug("expJournalAdd().get 메서드 expCode is {}",expCode);
+		//세션에서 아이디 값 확인
+		String loginId = (String)session.getAttribute("loginId");
+		logger.debug("expJournalAdd().post 메서드 loginId is {}",loginId);
+		//로그인 상태 확인
+		if(loginId==null) {
+			return "redirect:/member/login";
+		}
 		// 체험 일지 작성을 위해서 expCode 를 보냄
 		model.addAttribute("expCode", expCode);
 		return "/experience/expJournalAdd";
@@ -233,10 +314,17 @@ public class ExpController {
 	
 	// 체험 개인 정보 /experience/expInfo 을 get 방식으로 호출할때 발생
 	@RequestMapping(value = "/experience/expInfo", method = RequestMethod.GET)
-	public String ExpOne(Model model,@RequestParam(value="expCode") String expCode) {
+	public String ExpOne(Model model,@RequestParam(value="expCode") String expCode, HttpSession session) {
 		logger.debug("ExpController 호출 {ExpOne.get}.");
 		//넘어온 expCode 값 확인
 		logger.debug("ExpOne().get 메서드 expCode is {}",expCode);
+		//세션에서 아이디 값 확인
+		String loginId = (String)session.getAttribute("loginId");
+		logger.debug("ExpOne().post 메서드 loginId is {}",loginId);
+		//로그인 상태 확인
+		if(loginId==null) {
+			return "redirect:/member/login";
+		}
 		//expCode으로 체험 상세정보,체험일지 작성횟수 확인 받아옴
 		Map<String, Object> map = expService.selectExpOneInfo(expCode);
 		logger.debug("ExpOne().get 메서드 map is {}",map);
@@ -254,10 +342,17 @@ public class ExpController {
 	// 체험 등록 뷰 /experience/expAdd 을 get 방식으로 호출할때 실행
 	@RequestMapping(value = "/experience/expAdd", method = RequestMethod.GET)
 	public String expAdd(Model model,@RequestParam(value="animalCode") String animalCode,
-						@RequestParam(value="blCode") String blCode) {
+						@RequestParam(value="blCode") String blCode, HttpSession session) {
 		logger.debug("ExpController 호출 {expAdd.get}.");
 		//넘어온 animalCode 값 확인
 		logger.debug("expAdd().get 메서드 animalCode is {}",animalCode);
+		//세션에서 아이디 값 확인
+		String loginId = (String)session.getAttribute("loginId");
+		logger.debug("expAdd().post 메서드 loginId is {}",loginId);
+		//로그인 상태 확인
+		if(loginId==null) {
+			return "redirect:/member/login";
+		}
 		//model에 animalCode 담아 넘김
 		model.addAttribute("animalCode", animalCode);
 		//model에 blCode 담아 넘김
@@ -276,6 +371,10 @@ public class ExpController {
 		logger.debug("expAdd().post 메서드 exp is {}",exp);
 		String loginId = (String)session.getAttribute("loginId");
 		logger.debug("expAdd().post 메서드 loginId is {}",loginId);
+		//로그인 상태 확인
+		if(loginId==null) {
+			return "redirect:/member/login";
+		}
 		exp.setmExpId(loginId);
 		expService.addExp(exp);
 		return "redirect:/experience/expList";
